# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ParkWell - Parking Attendant System API",
    "description": "API documentation for the ParkWell system, for administrators and parking attendants.",
    "version": "1.1.0",
    "contact": {
      "name": "ParkWell Support",
      "email": "support@parkwell.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development Server (v1)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/vehicle-entries/enter": {
      "post": {
        "tags": [
          "Vehicle Operations (Attendant/Admin)"
        ],
        "summary": "Record vehicle entry",
        "operationId": "recordVehicleEntry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordVehicleEntryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle entry recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleEntryResponseWrapper"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/vehicle-entries/{entryId}/entry-ticket": {
      "parameters": [
        {
          "name": "entryId",
          "in": "path",
          "required": true,
          "description": "UUID of the Vehicle Entry",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "Public Tickets & Bills"
        ],
        "summary": "Download entry ticket PDF (Public)",
        "operationId": "downloadEntryTicket",
        "security": [],
        "responses": {
          "200": {
            "description": "PDF ticket file.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/vehicle-entries/{vehicleEntryId}/exit": {
      "parameters": [
        {
          "name": "vehicleEntryId",
          "in": "path",
          "required": true,
          "description": "UUID of the Vehicle Entry to exit",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "tags": [
          "Vehicle Operations (Attendant/Admin)"
        ],
        "summary": "Record vehicle exit",
        "operationId": "recordVehicleExit",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle exit recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleExitResponseWrapper"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/vehicle-entries/{entryId}/exit-bill": {
      "parameters": [
        {
          "name": "entryId",
          "in": "path",
          "required": true,
          "description": "UUID of the Vehicle Entry for the bill",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "Public Tickets & Bills"
        ],
        "summary": "Download exit bill PDF (Public)",
        "operationId": "downloadExitBill",
        "security": [],
        "responses": {
          "200": {
            "description": "PDF bill file.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Specific error message"
          }
        }
      },
...
      "VehicleEntryStatusEnum": {
        "type": "string",
        "enum": [
          "PARKED",
          "EXITED"
        ],
        "example": "PARKED"
      },
...
      "RecordVehicleEntryRequest": {
        "type": "object",
        "required": [
          "plate_number",
          "parking_id"
        ],
        "properties": {
          "plate_number": {
            "type": "string",
            "example": "ABC-123"
          },
          "parking_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "VehicleEntryDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plate_number": {
            "type": "string"
          },
          "entry_time": {
            "type": "string",
            "format": "date-time"
          },
          "exit_time": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "calculated_duration_minutes": {
            "type": "integer",
            "nullable": true
          },
          "charged_amount": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/VehicleEntryStatusEnum"
          },
          "ticket_number": {
            "type": "string"
          },
          "parking_id": {
            "type": "string",
            "format": "uuid"
          },
          "parking": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "charge_per_hour": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "recorded_by_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "recorded_by": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VehicleEntryResponseWrapper": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "vehicleEntry": {
            "$ref": "#/components/schemas/VehicleEntryDetails"
          },
          "ticketDownloadUrl": {
            "type": "string",
            "format": "url"
          }
        }
      },
      "VehicleExitResponseWrapper": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "vehicleEntry": {
            "$ref": "#/components/schemas/VehicleEntryDetails"
          },
          "billDownloadUrl": {
            "type": "string",
            "format": "url"
          },
          "parkingStatus": {
            "type": "string"
          }
        }
      },
...
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized. Token missing or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden. Insufficient permissions.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Bad Request. Invalid input or validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}